requirements.txt
transformers
torch
googlesearch-python